// web/src/api/generated/model/reportResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ReportCheck } from './reportCheck';
import type { ReportColumn } from './reportColumn';
import type { ReportRow } from './reportRow';

export interface ReportResult {
  key: string;
  titleEn: string;
  titleAr: string;
  subtitleEn: string;
  subtitleAr: string;
  companyNameEn: string;
  companyNameAr: string;
  currencyCode: string;
  minorUnits: number;
  columns: ReportColumn[];
  rows: ReportRow[];
  checks: ReportCheck[];
}