// web/src/api/generated/model/reportRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ReportCell } from './reportCell';
import type { ReportLink } from './reportLink';
import type { RowStyle } from './rowStyle';

export interface ReportRow {
  cells: ReportCell[];
  level?: number;
  style?: RowStyle;
  link?: null | ReportLink;
}